fit_gam fit a Generalized Additive Model to butterfly count data along a temporal variable and accounting for site effect when multiple are available.

fit_gam fit a Generalized Additive Model to butterfly count data along a temporal variable and accounting for site effect when multiple are available.

fit_gam(
  dataset_y,
  NbrSample = NULL,
  GamFamily = "poisson",
  MaxTrial = 4,
  SpeedGam = TRUE,
  OptiGam = TRUE,
  ConLikelihood = TRUE,
  TimeUnit = "d",
  MultiVisit = "mean"
)

Arguments

dataset_y
data.table with filtered butterfly counts for species x over year y over all sites.

NbrSample
integer inherited from flight_curve, default=100.

GamFamily
string inherited from flight_curve, default='poisson', but can be 'nb' or 'quasipoisson'.

MaxTrial
integer inherited from flight_curve, default=3.

SpeedGam
Logical to use the bam method instead of the gam method.

OptiGam
Logical to set use bam when data are larger than 100 and gam for smaller dataset

ConLikelihood
Logical to use the concentrated likelihood approach with not site parameters (default = TRUE).

TimeUnit
Character defining if the spline should be computed at the day 'd' or the week 'w'.

MultiVisit
function to be applied on multiple count within a single time unit, 'max' or 'mean' (default).

...
additional parameters passed to gam or bam function from the gam package.

Value

A list with three objects, i) **f_curve**: a data.table with the flight curve f_curve with expected relative abundance, normalize to sum to one over a full season, ii) **f_model**: the resulting gam model f_model fitted on the count data and iii) **f_data**: a data.table with the data used to fit the GAM model. This is provide for one year 'y'.

See also